refactor blocks
