+++ lib/blocks/widgets/block_slider/block_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter_annulus/shared/theme.dart';
 import 'package:flutter_annulus/shared/utils/nav_utils.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:responsive_framework/responsive_breakpoints.dart';
+import 'package:skeletonizer/skeletonizer.dart';
 
 import '../../../shared/providers/app_theme_provider.dart';
 import '../../../shared/utils/theme_color.dart';
@@ -11,14 +12,14 @@ import '../../models/block.dart';
 
 /// Block view widget
 class BlockView extends ConsumerWidget {
-  final AsyncValue<Block> asyncBlock;
+  final Block block;
 
   static Key blockItemKey(String blockId) => Key('blockItem-$blockId');
 
   // Constructor for BlockView widget
   const BlockView({
     Key? key,
-    required this.asyncBlock,
+    required this.block,
   }) : super(key: key);
 
   @override
@@ -39,89 +40,80 @@ class BlockView extends ConsumerWidget {
           width: 1.0,
         ),
       ),
-      child: asyncBlock.when(
-        error: (error, stackTrace) {
-          return const Center(
-            child: Text('Error'),
-          );
-        },
-        loading: () => const Center(child: CircularProgressIndicator()),
-        data: (Block block) {
-          return TextButton(
-              key: blockItemKey(block.header),
-              onPressed: () {
-                goToBlockDetails(
-                  context: context,
-                  block: block,
-                );
+      child: TextButton(
+          key: blockItemKey(block.header),
+          onPressed: () {
+            goToBlockDetails(
+              context: context,
+              block: block,
+            );
+          },
+          style: ButtonStyle(
+            foregroundColor: MaterialStateProperty.all<Color>(Colors.blue),
+            overlayColor: MaterialStateProperty.resolveWith<Color?>(
+              (Set<MaterialState> states) {
+                if (states.contains(MaterialState.hovered)) {
+                  return Colors.white.withOpacity(0.0);
+                }
+                if (states.contains(MaterialState.focused) || states.contains(MaterialState.pressed)) {
+                  return Colors.blue.withOpacity(0.12);
+                }
+                return null; // Defer to the widget's default.
               },
-              style: ButtonStyle(
-                foregroundColor: MaterialStateProperty.all<Color>(Colors.blue),
-                overlayColor: MaterialStateProperty.resolveWith<Color?>(
-                  (Set<MaterialState> states) {
-                    if (states.contains(MaterialState.hovered)) {
-                      return Colors.white.withOpacity(0.0);
-                    }
-                    if (states.contains(MaterialState.focused) || states.contains(MaterialState.pressed)) {
-                      return Colors.blue.withOpacity(0.12);
-                    }
-                    return null; // Defer to the widget's default.
-                  },
+            ),
+          ),
+          child: Container(
+            padding: const EdgeInsets.all(40.0),
+            width: 240,
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Skeleton.keep(
+                  child: SizedBox(
+                    child: Image.asset(
+                      'assets/icons/block-icon.png',
+                      width: 20,
+                      height: 20,
+                    ),
+                  ),
                 ),
-              ),
-              child: Container(
-                  padding: const EdgeInsets.all(40.0),
-                  width: 240,
-                  child: Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Container(
-                          width: 45,
-                          height: 45,
-                          decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(16.0),
-                              border: Border.all(
-                                  color: getSelectedColor(colorTheme, 0xFFE7E8E8, 0xFF4B4B4B),
-                                  style: BorderStyle.solid,
-                                  width: 1.0)),
-                          child: Image.asset(
-                            colorTheme == ThemeMode.light
-                                ? 'assets/icons/block-icon.png'
-                                : 'assets/icons/dark-block-icon.png',
-                            width: 20,
-                            height: 20,
-                          ),
-                        ),
-                        const SizedBox(
-                          height: 16,
-                        ),
-                        Text(block.height.toString(), style: titleLarge(context)),
-                        BlockSmallText(text: blockTimeStampToDifference(block)),
-                        const SizedBox(
-                          height: 8,
-                        ),
-                        BlockHeaderText(text: block.epoch.toString()),
-                        const BlockSmallText(text: 'Epoch'),
-                        const SizedBox(
-                          height: 8,
-                        ),
-                        BlockHeaderText(text: block.header),
-                        const BlockSmallText(text: 'Header'),
-                        const SizedBox(
-                          height: 8,
-                        ),
-                        Expanded(
-                          child: BlockHeaderText(
-                            text: DateTime.fromMicrosecondsSinceEpoch(block.timestamp).toString().substring(11, 19),
-                          ),
-                        ),
-                        const Expanded(child: BlockSmallText(text: 'UTC')),
-                      ]))
-              //cardChild,
-              );
-        },
-      ),
+                const SizedBox(
+                  height: 16,
+                ),
+                Text(
+                  block.height.toString(),
+                  style: titleLarge(context),
+                ),
+                BlockSmallText(
+                  text: blockTimeStampToDifference(block),
+                ),
+                const SizedBox(
+                  height: 8,
+                ),
+                BlockHeaderText(text: block.epoch.toString()),
+                const BlockSmallText(text: 'Epoch'),
+                const SizedBox(
+                  height: 8,
+                ),
+                BlockHeaderText(text: block.header),
+                const BlockSmallText(text: 'Header'),
+                const SizedBox(
+                  height: 8,
+                ),
+                Expanded(
+                  child: BlockHeaderText(
+                    text: DateTime.fromMicrosecondsSinceEpoch(block.timestamp).toString().substring(11, 19),
+                  ),
+                ),
+                const Expanded(
+                  child: BlockSmallText(text: 'UTC'),
+                ),
+              ],
+            ),
+          )
+          //cardChild,
+          ),
     );
   }
 }
