+++ lib/home/screen/home_screen.dart
@@ -81,7 +81,7 @@ class HomeScreen extends HookConsumerWidget {
                 ],
               ),
             ),
-            BlockViewSlider(),
+            const BlockViewSlider(),
             const Transactions(),
           ],
         ),
